fix(MatterEngine3): world_tracer stale-hit commit, 64-bit bounds checks; refactor: shared mat_math.h, j_ts_layer helper
B7: world_tracer.cpp — only commit best_t/normal/material when tri_idx is valid; previously best_t advanced with stale normal/material on out-of-range BVH prim indices.

B12: tileset_gtex.cpp — cast uint32 offset+size to uint64_t before bounds comparison; the 32-bit sum could wrap and bypass the guard on >2GB-offset blobs.

mat_math.h: create internal header holding mul16 (row-major 4x4 multiply) and NormalMat (inverse-transpose upper-3x3 for normals); remove per-file copies from world_tracer.cpp and part_flatten.cpp. mat_invert/mat_mul (raylib Matrix), mat4_mul (mat4 struct), and the Shepperd extraction functions are semantically different across files — kept local as noted in report.

tileset_settle.cpp: replace raw new/delete Impl with std::unique_ptr<Impl>.

dsl_bindings.cpp: extract place_one_instance() static helper from j_ts_layer's two near-identical per-point placement blocks (strips + interior). Bake outputs byte-identical (golden tests pass).

+++ MatterEngine3/src/tileset_gtex.cpp
@@ -248,7 +248,9 @@ bool load_gtex(const std::string& path,
     // Decode PNG blobs (RGB8/RG8/RGB8) via stb_image; R16 is raw uint16 LE.
     auto decode_png = [&](int chan_id, int expect_comp, std::vector<uint8_t>& out) -> bool {
         const GTexChannelEntry* e = by_id[chan_id];
-        if (e->offset + e->size > raw.size()) return false;
+        // Cast to uint64_t before adding: both fields are uint32_t and their sum
+        // can wrap to a value that passes the bounds check on a >2 GB file (B12).
+        if ((uint64_t)e->offset + (uint64_t)e->size > raw.size()) return false;
         int w = 0, h = 0, comp = 0;
         stbi_uc* px = stbi_load_from_memory(raw.data() + e->offset, (int)e->size,
                                             &w, &h, &comp, expect_comp);
@@ -270,7 +272,8 @@ bool load_gtex(const std::string& path,
     // Height R16 raw.
     {
         const GTexChannelEntry* e = by_id[CHAN_HEIGHT_R16];
-        if (e->offset + e->size > raw.size()) {
+        // Cast to uint64_t before adding to avoid uint32 wrap (B12).
+        if ((uint64_t)e->offset + (uint64_t)e->size > raw.size()) {
             err = "load_gtex: height blob overflow: " + path; return false;
         }
         if (e->size != e->width * e->height * 2) {
